LACosmic.jl

)

Documentation for LACosmic.

LACosmic.subsample

LACosmic.subsample — Function

LACosmic.subsample(array, block_size=2)

Subsample array by the given factor without copying or allocating. This effectively treats each pixel as block_size x block_size pixels. The value of each pixel is not normalized or averaged. This is a view into the parent array, and if the data was changed this view would change subsequently.

Examples

julia> arr = [1 2; 3 4]

julia> sub = LACosmic.subsample(arr)
4×4 LACosmic.SubsampledArray{Int64, 2, Matrix{Int64}}:
 1  1  2  2
 1  1  2  2
 3  3  4  4
 3  3  4  4

julia> all(sub[1:2, 1:2] .=== arr[1, 1])
true

julia> sub[3, 3] = 5;

julia> sub
4×4 LACosmic.SubsampledArray{Int64, 2, Matrix{Int64}}:
 1  1  2  2
 1  1  2  2
 3  3  5  5
 3  3  5  5